feat(portal): configure email sent at end of have your say journey
