+++ apps/portal/src/app/config-types.d.ts
@@ -22,8 +22,9 @@ interface Config {
 	govNotify: {
 		disabled: boolean;
 		apiKey: string;
-		templates: {
+		templateIds: {
 			test: string;
+			acknowledgementOfRepresentation: string;
 		};
 	};
 	crownDevContactInfo: {
