+++ apps/portal/src/app/config.js
@@ -35,6 +35,7 @@ export function loadConfig() {
 		GOV_NOTIFY_DISABLED,
 		GOV_NOTIFY_API_KEY,
 		GOV_NOTIFY_TEST_TEMPLATE_ID,
+		GOV_NOTIFY_ACK_REP_TEMPLATE_ID,
 		CROWN_DEV_CONTACT_EMAIL
 	} = process.env;
 
@@ -55,7 +56,11 @@ export function loadConfig() {
 
 	const govNotifyDisabled = GOV_NOTIFY_DISABLED === 'true';
 	if (!govNotifyDisabled) {
-		const props = { GOV_NOTIFY_API_KEY, GOV_NOTIFY_TEST_TEMPLATE_ID };
+		const props = {
+			GOV_NOTIFY_API_KEY,
+			GOV_NOTIFY_TEST_TEMPLATE_ID,
+			GOV_NOTIFY_ACK_REP_TEMPLATE_ID
+		};
 		for (const [k, v] of Object.entries(props)) {
 			if (v === undefined || v === '') {
 				throw new Error(k + ' must be a non-empty string');
@@ -102,8 +107,9 @@ export function loadConfig() {
 		govNotify: {
 			disabled: govNotifyDisabled,
 			apiKey: GOV_NOTIFY_API_KEY,
-			templates: {
-				test: GOV_NOTIFY_TEST_TEMPLATE_ID
+			templateIds: {
+				test: GOV_NOTIFY_TEST_TEMPLATE_ID,
+				acknowledgementOfRepresentation: GOV_NOTIFY_ACK_REP_TEMPLATE_ID
 			}
 		},
 		crownDevContactInfo: {
