+++ infrastructure/app-portal.tf
@@ -76,9 +76,10 @@ module "app_portal" {
     WEBSITE_AUTH_AAD_ALLOWED_TENANTS         = data.azurerm_client_config.current.tenant_id
 
     # gov notify
-    GOV_NOTIFY_DISABLED         = var.apps_config.gov_notify.disabled
-    GOV_NOTIFY_API_KEY          = local.key_vault_refs["crown-gov-notify-api-key"]
-    GOV_NOTIFY_TEST_TEMPLATE_ID = var.apps_config.gov_notify.templates.test_template_id
+    GOV_NOTIFY_DISABLED            = var.apps_config.gov_notify.disabled
+    GOV_NOTIFY_API_KEY             = local.key_vault_refs["crown-gov-notify-api-key"]
+    GOV_NOTIFY_TEST_TEMPLATE_ID    = var.apps_config.gov_notify.templates.test_template_id
+    GOV_NOTIFY_ACK_REP_TEMPLATE_ID = var.apps_config.gov_notify.templates.ack_rep_template_id
 
     CROWN_DEV_CONTACT_EMAIL = var.apps_config.contact_email
