+++ infrastructure/environments/dev.tfvars
@@ -43,6 +43,7 @@ apps_config = {
     templates = {
       test_template_id    = "4b8adfb1-1b7c-4333-b512-761eeedfdca2"
       pre_ack_template_id = "298a986d-c142-46f5-804a-9cb853ba8b3d"
+      ack_rep_template_id = "52312d8e-2af8-4212-a280-bfc6106cc56d"
     }
   }
