+++ infrastructure/variables.tf
@@ -51,6 +51,7 @@ variable "apps_config" {
       templates = object({
         test_template_id    = string
         pre_ack_template_id = string
+        ack_rep_template_id = string
       })
     })
