+++ packages/lib/govnotify/gov-notify-client.js
@@ -31,6 +31,18 @@ export class GovNotifyClient {
 	}
 
 	/**
+	 * @param {string} email - Recipients email address
+	 * @param {import('./types.js').SendAcknowledgementOfRepresentationPersonalisation} personalisation
+	 * @returns {Promise<void>}
+	 */
+	async sendAcknowledgementOfRepresentation(email, personalisation) {
+		console.log(this.#templateIds);
+		await this.sendEmail(this.#templateIds.acknowledgementOfRepresentation, email, {
+			personalisation: personalisation
+		});
+	}
+
+	/**
 	 * @param {string} templateId - Gov Notify email template id
 	 * @param {string} emailAddress - Recipients email address
 	 * @param {GovNotifyOptions} options - Options to pass to Gov Notify
